Add "bind" port configuration for Serve command.
The "bind" config allows us to instruct serve to bind onto a different
address than the "host", so we can attach to an interface other than "localhost"
which is a loopback interface (ie. not public).

The "host" setting defines the URL when running in "develop" mode, making it
inappropriate to set to anything but the "public" hostname of the machine serving files.

When running locally, both "bind" and "host" would be "localhost", and the
connection is isolated to your local machine.
However, when using something like a Docker container the localhost binding prevents
access to served content. In this instance "bind" needs to be "0.0.0.0"
(for a public if connection) and "host" needs to be the ip/hostname of the container.

For backwards compatibility, if "bind" is not provided, it falls back to
the value defined by "host".